+++ wilhelm_renderer_sys/build.rs
@@ -60,12 +60,6 @@ fn main() {
         );
         println!("cargo:rustc-link-lib=static=glrenderer");
         println!("cargo:rustc-link-lib=static=glfw3");
-        // FreeType uses 'd' suffix for debug builds
-        if profile == "debug" {
-            println!("cargo:rustc-link-lib=static=freetyped");
-        } else {
-            println!("cargo:rustc-link-lib=static=freetype");
-        }
 
         if link_gl {
             println!("cargo:rustc-link-lib=dylib=GL");
@@ -78,12 +72,6 @@ fn main() {
         );
         println!("cargo:rustc-link-lib=static=glrenderer");
         println!("cargo:rustc-link-lib=static=glfw3");
-        // FreeType uses 'd' suffix for debug builds
-        if profile == "debug" {
-            println!("cargo:rustc-link-lib=static=freetyped");
-        } else {
-            println!("cargo:rustc-link-lib=static=freetype");
-        }
 
         println!("cargo:rustc-link-lib=framework=CoreFoundation");
         println!("cargo:rustc-link-lib=framework=IOKit");
@@ -101,12 +89,6 @@ fn main() {
         println!("cargo:rustc-link-search=native={}", build_dir.display());
         println!("cargo:rustc-link-lib=static=glrenderer");
         println!("cargo:rustc-link-lib=static=glfw3");
-        // FreeType uses 'd' suffix for debug builds
-        if profile == "debug" {
-            println!("cargo:rustc-link-lib=static=freetyped");
-        } else {
-            println!("cargo:rustc-link-lib=static=freetype");
-        }
 
         // Link Windows system libraries
         println!("cargo:rustc-link-lib=dylib=opengl32");